Manually call `to_s` on all URIs
It looks like this change was introduced in sferik/twitter@341f68d1a

The Twitter gem now returns URIs instead of strings, so we have to
manually call to_s when we want a string since we are using Object
dumping to convert into a JSON blob.

This should fix the issues in mroth/emojitrack-feeder#3, but more
testing is required.