use naked attributes to get raw strings
This lets us get at some raw strings for things that we are going to
convert back to string anyhow — the convenience accessors allocate
native Ruby objects, AND memoize them.  However, they do this on first
access, so by avoiding calling them we can avoid unneeded processing
and memory allocation.

The exception is the created_at timestamp, which appears to be parsed
and memorized at Tweet object instantiation.  This is fine, since we
need to convert to ISO-8601 anyhow, so we continue using the accessor
method rather than parsing on our own for a second time (which
negatives our speed games, parsing time is slow!)