remove memoization
trying to get some more data off the heap and avoid extra allocation.

instead we will try to be careful about not calling this stuff more
than once (removing kiosk mode should help).

be explicit when calling unmemoized functions in feeder so we don’t
think of them as attributes.